Updating README.md w/ valid install command
The readme appears to contain commands to install this plugin for an older version of dokku. Since the readme mentions requiring 0.4.0, I've updated it to show the commands which work for installing this plugin. Someone more familiar with dokku should review how the plugin:install-dependencies is handled here, as it could be incorrect.